+++ CHANGELOG.md
@@ -1,15 +1,24 @@
 # Changelog
 
-## 0.1.2
+
+## v0.1.4
+
+* Updated to fyne 2.3.5
+* Added extra `pkg-config` used for X11.
+
+## v0.1.3
+(Skip)
+
+## v0.1.2
 
 * Added "screenshot in 5 seconds" option to systray.
 * Added "Delayed Screenshot" menu option in GoShot, with arbitrary delay.
 
-## 0.1.1
+## v0.1.1
 
 * Added -ldconfig="-H windowsgui" for windows to removed annoying terminal window.
 * Saved font size in preferences.
 * Added `goshot.syso` with the icon to embed in the Windows version.
 
-## 0.1.0
+## v0.1.0
 Initial release: windows, linux